interrim commit

+++ src/dose/dose_dobexplorer.ss/src/graphwidget.cpp
@@ -23,8 +23,11 @@
 ******************************************************************************/
 
 #include "graphwidget.h"
-//#include <iostream>
 #include <algorithm>
+#include <QPaintEvent>
+#include <QDateTime>
+#include <QPainter>
+#include <QPainterPath>
 
 GraphWidget::GraphWidget(QWidget * parent):
     QFrame(parent),
@@ -55,11 +58,11 @@ void GraphWidget::paintEvent(QPaintEvent* event)
 
     QPainterPath path;
 
-    path.moveTo(start.secsTo(m_data.begin()->first), m_data.begin()->second);
+    path.moveTo(static_cast<qreal>(start.secsTo(m_data.begin()->first)), m_data.begin()->second);
     for (PlotData::iterator it = ++m_data.begin();
          it != m_data.end(); ++it)
     {
-        path.lineTo(start.secsTo(it->first),it->second);
+        path.lineTo(static_cast<qreal>(start.secsTo(it->first)),it->second);
     }
 
     painter.drawPath(path);
